Add Groups
Groups are a way to store a search on the entity pool for later.

Internally they store a query expression, a local cache and a
reference to the pools entity store, when an item is added or removed
that would have/does match the query, the cache is invalidated and
refreshed the next time the group's contents are asked for.